+++ packages/paraglide/paraglide-next/example/next.config.mjs
@@ -1,0 +1,13 @@
+import { withParaglideNext } from "@inlang/paraglide-next";
+
+export default withParaglideNext({
+	paraglide: {
+		outdir: "./src/paraglide",
+		project: "./project.inlang",
+	},
+	eslint: {
+		// Warning: This allows production builds to successfully complete even if
+		// your project has ESLint errors.
+		ignoreDuringBuilds: true,
+	},
+});