+++ packages/paraglide/paraglide-sveltekit/example/vite.config.ts
@@ -1,0 +1,13 @@
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vite';
+import { paraglideSveltekit } from '@inlang/paraglide-sveltekit';
+
+export default defineConfig({
+	plugins: [
+		sveltekit(),
+		paraglideSveltekit({
+			project: './project.inlang',
+			outdir: './src/lib/paraglide'
+		})
+	]
+});