+++ packages/plugins/inlang-message-format/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @inlang/plugin-message-format
 
+## 3.1.1
+
+### Patch Changes
+
+- 8132942: feat: support array of paths for pathPattern in inlang-message-format plugin
+
+  ```diff
+  // settings.json
+
+  {
+    "plugin.inlang.messageFormat": {
+  +    pathPattern: ["/defaults/{locale}.json", "/translations/{locale}.json"],
+    }
+  }
+
+  ```
+
+## 3.1.0
+
+### Minor Changes
+
+- 4adfd4d: re-enables adding the `$schema` key to exported files
+
+  the `$schema` prop enables IDEs to provide autocompletion and type checking for the message files
+
+### Patch Changes
+
+- 997f55b: update the fileschema for to variants
+
+  closes https://github.com/opral/inlang-paraglide-js/issues/319
+
 ## 3.0.3
 
 ### Patch Changes